Implement DiceDistributionChart
