+++ src/GameStat/index.tsx
@@ -19,7 +19,10 @@ import { GameID, useQueryGame } from '../api'
 
 import { TotalDurationChart } from './TotalDurationChart'
 import { TurnsDistributionChart } from './TurnsDistributionChart'
-import { NumberDiceDistribution } from './NumberDiceDistribution'
+import {
+  NumberDiceDistribution,
+  EventDieDistribution
+} from './DiceDistributionChart'
 
 ChartJS.register(
   ArcElement,
@@ -62,6 +65,7 @@ export const GameStat: React.VFC = React.memo(() => {
       <TotalDurationChart game={game} />
       <TurnsDistributionChart turns={game.turns} />
       <NumberDiceDistribution turns={game.turns} />
+      <EventDieDistribution turns={game.turns} />
     </ViewContainer>
   )
 })
